Extract input component rendering to another component
