+++ src/routes/safe/components/Balances/SendModal/screens/ContractInteraction/RenderInputParams/index.tsx
@@ -1,12 +1,9 @@
 import React from 'react'
 import { useField } from 'react-final-form'
 
-import Field from 'src/components/forms/Field'
-import TextField from 'src/components/forms/TextField'
-import { composeValidators, mustBeEthereumAddress, required } from 'src/components/forms/validator'
-import Col from 'src/components/layout/Col'
 import Row from 'src/components/layout/Row'
-import Checkbox from 'src/components/forms/Checkbox'
+
+import InputComponent from './InputComponent'
 
 const RenderInputParams = () => {
   const {
@@ -17,54 +14,6 @@ const RenderInputParams = () => {
   }: any = useField('selectedMethod', { value: true })
   const renderInputs = validABI && !!method && method.inputs.length
 
-  const renderInputComponent = (type, name, key, placeholder) => {
-    if (!type) {
-      return null
-    }
-    switch (type) {
-      case 'bool': {
-        const inputProps = {
-          'data-testid': key,
-        }
-        return (
-          <Col>
-            <Field component={Checkbox} name={key} label={placeholder} type="checkbox" inputProps={inputProps} />
-          </Col>
-        )
-      }
-      case 'address': {
-        return (
-          <Col>
-            <Field
-              component={TextField}
-              name={key}
-              placeholder={placeholder}
-              testId={key}
-              text={placeholder}
-              type="text"
-              validate={composeValidators(required, mustBeEthereumAddress)}
-            />
-          </Col>
-        )
-      }
-      default: {
-        return (
-          <Col>
-            <Field
-              component={TextField}
-              name={key}
-              placeholder={placeholder}
-              testId={key}
-              text={placeholder}
-              type="text"
-              validate={required}
-            />
-          </Col>
-        )
-      }
-    }
-  }
-
   return !renderInputs
     ? null
     : method.inputs.map(({ name, type }, index) => {
@@ -73,7 +22,7 @@ const RenderInputParams = () => {
 
         return (
           <Row key={key} margin="sm">
-            {renderInputComponent(type, name, key, placeholder)}
+            <InputComponent type={type} keyValue={key} placeholder={placeholder} />
           </Row>
         )
       })
